Move CreateDataSourceTable and CreateDataSourceTableAsSelect to sql/core.
